pyproject.toml
[tool.poetry]
name = "grammargrove"
version = "0.1.0"
description = ""
authors = ["Andrew Deaver <deaver.andrew@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.0.0"
psycopg2 = "^2.9.6"
uwsgi = "^2.0.21"
django-extensions = "^3.2.3"
boto3 = "^1.26.151"
openai = "^0.27.8"
jieba = "^0.42.1"
pygtrie = "^2.5.0"
stripe = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
